add bluetooth flyout. fix multi-device displaying
Added the Windows 11 bluetooth quick settings flyout as an item to the context menu. This has some annoyances like clicking out of it isn't the smoothest, but this is a Windows 11 limitation, and I don't think its worth hacking a fix at the moment.

Fixed multi-device displaying so the text doesn't all render on top of the bottom device, also enhanced it a bit with arrows showing the direction of the apps in each device as a queue to the user to read bottom-up