Rework flyout to be "bottom up"
its better UX if what you're most likely to control is as close to where the mouse is going to be as possible.

added a little visual arrows when multiple endpoints are active to show the direction of the menu / app submenus.

also made the collapse / expand button always visible since this state is saved across flyout opens. Otherwise you lose the ability to set this when you don't have multiple endpoints active.

updated the Windows SDK versions.